.github/skills/skill-creator/requirements.txt
# Core dependencies
pyyaml>=6.0

# LLM backends (install the one you need)
openai>=1.0.0
anthropic>=0.30.0

